# Conflicts:
#	src/main/webapp/WEB-INF/views/realestate/realestateList.jsp

+++ src/main/java/com/project/restaurant/vo/ResHashtag.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ResHashtag {
     private String resNo;
     private String hashtagId;
-    private int revNo;
+    private long revNo;
     private long memberNo;
 
     public ResHashtag() {}
